arrange task manager code
