+++ arch/task_manager/example_conf/query_model_version.json
@@ -1,0 +1,5 @@
+{
+    "task": "model_version",
+    "work_mode": 0,
+    "namespace": "hetero_lr"
+}
