+++ arch/task_manager/utils/grpc_utils.py
@@ -1,0 +1,70 @@
+#
+#  Copyright 2019 The FATE Authors. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+import requests
+import json
+from arch.api.proto import proxy_pb2, proxy_pb2_grpc
+from arch.api.proto import basic_meta_pb2
+import grpc
+from arch.task_manager.settings import ROLE, IP, GRPC_PORT, LOCAL_URL, PROXY_HOST, PROXY_PORT, \
+    PARTY_ID, HEADERS
+
+
+def get_proxy_data_channel():
+    channel = grpc.insecure_channel('{}:{}'.format(PROXY_HOST, PROXY_PORT))
+    stub = proxy_pb2_grpc.DataTransferServiceStub(channel)
+    return channel, stub
+
+
+def wrap_grpc_packet(_json_body, _method, _url, _dst_party_id=None, job_id=None):
+    _src_end_point = basic_meta_pb2.Endpoint(ip=IP, port=GRPC_PORT)
+    _src = proxy_pb2.Topic(name=job_id, partyId="{}".format(PARTY_ID), role=ROLE, callback=_src_end_point)
+    _dst = proxy_pb2.Topic(name=job_id, partyId="{}".format(_dst_party_id), role=ROLE, callback=None)
+    _task = proxy_pb2.Task(taskId=job_id)
+    _command = proxy_pb2.Command(name=ROLE)
+    _meta = proxy_pb2.Metadata(src=_src, dst=_dst, task=_task, command=_command, operator=_method)
+    _data = proxy_pb2.Data(key=_url, value=bytes(json.dumps(_json_body), 'utf-8'))
+    return proxy_pb2.Packet(header=_meta, body=_data)
+
+def get_url(_suffix):
+    return "{}{}".format(LOCAL_URL, _suffix)
+
+class UnaryServicer(proxy_pb2_grpc.DataTransferServiceServicer):
+
+
+    def unaryCall(self, _request, context):
+        packet = _request
+        #manager.logger.info("<=== RPC receive: {}".format(packet))
+        header = packet.header
+        _suffix = packet.body.key
+        param_bytes = packet.body.value
+        param = bytes.decode(param_bytes)
+        job_id = header.task.taskId
+        dst = header.src
+        method = header.operator
+        #manager.logger.info("From : {} \tjob_id : {}\tmethod : {}\tparams: {}".format(dst, job_id, method, param))
+
+        action = getattr(requests, method.lower(), None)
+        if action:
+            resp = action(url=get_url(_suffix), data=param, headers=HEADERS)
+        else:
+            pass
+            #manager.logger.error("method {} not supported".format(method))
+
+        resp_json = resp.json()
+
+        #manager.logger.info("Local http response: {}".format(resp_json))
+
+        return wrap_grpc_packet(resp_json, method, _suffix, dst.partyId, job_id)
